backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
pg8000>=1.30.0
pymongo==4.6.1
python-dotenv==1.0.0
Werkzeug==3.0.1